installer_stub: Added the ability to skip the installation of the SDE software.
make_vm_image:
Added the ability to skip the installation of the SDE software.
Changed the manner in which virt-sparsify is invoked in order to
workaround an apparent bug.

minimal_xfce_sde.ks:
Changed the lightdm package to the lightdm-autologin-greeter package to
avoid pulling in the deepin desktop.
Added the openssl-devel package.